Working Trade comments
Fixed comments, they will now update int he db to the specific trade offer. 
Both parties can now comment on the tradeoffers.
Fixed view for when the trade has been claimed to be completed. It will not inform the user the trade is done and to look at the block for more.